added poster type dropdown
